Writing handleSelect function
handleSelect access the navigation object, use method 'navigate' that takes 2 arguments ( 1.we use the screen we want to navigate to which is 'ModuleViewScreen' 2. we pass the information to the screen we want to reach. The second argument is an object. )
If you have the syntax { module: module } where you have the property name module and the variable that it has the same name you can collapse it down to the single word { module } and javaScript will know what you mean.