2.	Pass module information to ModuleViewScreen
In ModuleViewScreen we want to get the data from ModuleListScreen.
Every screen that is visited has navigation object and route object. The information that we want to pass exist in route object.
In the initialisations area we deconstruct the information from the route object that we need using params (in params information lives)