Dealing with a warning: “Non-serializable values …”
It is caused when we try to pass a callback function ‘onDelete’ via navigate. It doesn’t mind when the object is passed but it does not like when the function is being passed. It is caused by the fact that you cannot take the function and serialise it (store it on the disc).
We are going to use LogBox to ignore the warning message.