Getting rid of Add naming on the button and instead creating submitLabel variable
We dont want Add button to be called Add as we dont want hardcoded as Add because we might be using this form also for modifying not only adding. We create temporary variables "submitIcon" and "submitLabel" and because they have something to do with a view we will define them in the view section.